paper-agent/requirements.txt
requests==2.31.0
google-generativeai==0.3.2
notion-client==2.2.1
python-dotenv==1.0.0
schedule==1.2.0
feedparser==6.0.10
beautifulsoup4==4.12.2
arxiv==2.1.0
pydantic==2.5.0
loguru==0.7.2
